Change overriding approach
