--- app/views/decidim/accountability/admin/results/_form.html.erb
@@ -1,81 +1,0 @@
-<div class="card">
-  <div class="card-divider">
-    <h2 class="card-title"><%= title %></h2>
-  </div>
-
-  <div class="card-section">
-    <div class="row column">
-      <%= form.translated :text_field, :title, autofocus: true %>
-    </div>
-
-    <div class="row column">
-      <%= form.translated :editor, :description %>
-    </div>
-
-    <% if @form.parent_id %>
-
-      <div class="row column">
-        <%= form.select :parent_id, parent_results.map{|result| [translated_attribute(result.title), result.id] }, include_blank: true %>
-      </div>
-
-    <% else %>
-
-      <% if current_participatory_space.has_subscopes? %>
-        <div class="row column">
-          <%= scopes_picker_field form, :decidim_scope_id %>
-        </div>
-      <% end %>
-
-      <div class="row column">
-        <%= form.categories_select :decidim_category_id, current_participatory_space.categories, include_blank: true, disable_parents: false %>
-      </div>
-
-    <% end %>
-
-    <div class="row">
-      <div class="columns xlarge-6">
-        <%= form.date_field :start_date %>
-      </div>
-
-      <div class="columns xlarge-6">
-        <%= form.date_field :end_date %>
-      </div>
-    </div>
-
-    <div class="row">
-      <div class="columns xlarge-6">
-        <%= form.select :decidim_accountability_status_id, statuses.map{|status| [translated_attribute(status.name), status.id, { 'data-progress' => status.progress }] }, include_blank: true %>
-      </div>
-
-      <div class="columns xlarge-6">
-        <%= form.number_field :progress %>
-      </div>
-    </div>
-
-    <div class="row column">
-      <% if @form.proposals %>
-        <%= form.select :proposal_ids,
-                        @form.proposals,
-                        {},
-                        { multiple: true, class: "chosen-select" }
-        %>
-      <% end %>
-    </div>
-
-    <div class="row column">
-      <% if @form.projects %>
-        <%= form.select :project_ids,
-                        @form.projects,
-                        {},
-                        { multiple: true, class: "chosen-select" }
-        %>
-      <% end %>
-    </div>
-
-    <div class="row column">
-      <%= form.text_field :external_id %>
-    </div>
-
-  </div>
-</div>
-
